SUBIDA 2

+++ PARCIAL/src/Calculos.h
@@ -8,7 +8,7 @@
 #ifndef CALCULOS_H_
 #define CALCULOS_H_
 
-int dataCollection(datosCenso* datos, int lenDatosCenso);
+float dataCollection(datosCenso* datos, int lenDatosCenso);
 float promediar(int valorPivot, int valorPromediar);
 
 #endif /* CALCULOS_H_ */
